+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Remove `Transport.onMessageProcessed()` as flow control is no longer part of the application layer and is instead
   managed by the `WebSocket` implementation.
 * Removed sequences from the server-to-client message as messages can never be transmitted out of sequence.
+* Refactor `Connector.state` to be readable outside a transaction.
 
 ### [v6.37](https://github.com/replicant4j/replicant/tree/v6.37) (2019-03-18)
 [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.36...v6.37)